LPS-52768 Moves classes from assetpublisher to asset
