+++ portal-impl/src/com/liferay/portlet/assetpublisher/util/CustomUserAttributesAssetEntryQueryProcessor.java
@@ -20,6 +20,7 @@ import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.model.User;
 import com.liferay.portlet.asset.service.persistence.AssetEntryQuery;
+import com.liferay.portlet.asset.util.AssetEntryQueryProcessor;
 
 import javax.portlet.PortletPreferences;
