+++ portal-service/src/com/liferay/portlet/assetpublisher/util/AssetPublisherUtil.java
@@ -27,6 +27,7 @@ import com.liferay.portlet.asset.model.AssetEntry;
 import com.liferay.portlet.asset.model.AssetRendererFactory;
 import com.liferay.portlet.asset.model.ClassType;
 import com.liferay.portlet.asset.service.persistence.AssetEntryQuery;
+import com.liferay.portlet.asset.util.AssetEntryQueryProcessor;
 
 import java.util.List;
 import java.util.Locale;
